Introduce line break tests
