+++ src/Avalonia.Visuals/Media/TextFormatting/Unicode/LineBreakEnumerator.cs
@@ -95,6 +95,7 @@ namespace Avalonia.Media.TextFormatting.Unicode
 
                     if (_nextClass.Value == LineBreakClass.MandatoryBreak)
                     {
+                        _lastPos = _pos;
                         Current = new LineBreak(FindPriorNonWhitespace(_lastPos), _lastPos);
                         return true;
                     }
@@ -108,6 +109,7 @@ namespace Avalonia.Media.TextFormatting.Unicode
                 {
                     case PairBreakType.DI: // Direct break
                         shouldBreak = true;
+                        _lastPos = _pos;
                         break;
 
                     case PairBreakType.IN: // possible indirect break
