+++ tests/Avalonia.Visuals.UnitTests/Media/TextFormatting/BreakPairTable.txt
@@ -1,7 +1,7 @@
 	OP	CL	CP	QU	GL	NS	EX	SY	IS	PR	PO	NU	AL	HL	ID	IN	HY	BA	BB	B2	ZW	CM	WJ	H2	H3	JL	JV	JT	RI	EB	EM	ZWJ	CB
 OP	^	^	^	^	^	^	^	^	^	^	^	^	^	^	^	^	^	^	^	^	^	@	^	^	^	^	^	^	^	^	^	^	^
-CL	_	^	^	%	%	^	^	^	^	%	%	_	_	_	_	_	%	%	_	_	^	#	^	_	_	_	_	_	_	_	_	%	_
-CP	_	^	^	%	%	^	^	^	^	%	%	%	%	%	_	_	%	%	_	_	^	#	^	_	_	_	_	_	_	_	_	%	_
+CL	_	^	^	%	%	^	^	^	^	%	%	_	_	_	_	^	%	%	_	_	^	#	^	_	_	_	_	_	_	_	_	%	_
+CP	_	^	^	%	%	^	^	^	^	%	%	%	%	%	_	^	%	%	_	_	^	#	^	_	_	_	_	_	_	_	_	%	_
 QU	^	^	^	%	%	%	^	^	^	%	%	%	%	%	%	%	%	%	%	%	^	#	^	%	%	%	%	%	%	%	%	%	%
 GL	%	^	^	%	%	%	^	^	^	%	%	%	%	%	%	%	%	%	%	%	^	#	^	%	%	%	%	%	%	%	%	%	%
 NS	_	^	^	%	%	%	^	^	^	_	_	_	_	_	_	_	%	%	_	_	^	#	^	_	_	_	_	_	_	_	_	%	_
