+++ tests/Avalonia.Visuals.UnitTests/Media/TextFormatting/UnicodeDataGeneratorTests.cs
@@ -1,4 +1,6 @@
-﻿using Xunit;
+﻿using System;
+using Avalonia.Media.TextFormatting.Unicode;
+using Xunit;
 
 namespace Avalonia.Visuals.UnitTests.Media.TextFormatting
 {
@@ -12,6 +14,27 @@ namespace Avalonia.Visuals.UnitTests.Media.TextFormatting
         public void Should_Generate_Data()
         {
             UnicodeDataGenerator.Execute();
+        }
+        [Theory/*(Skip = "Only run when we update the trie.")*/]
+        [ClassData(typeof(LineBreakTestDataGenerator))]
+
+        public void Should_Enumerate_LineBreaks(string text, int expectedLength)
+        {
+            var textMemory = text.AsMemory();
+
+            var enumerator = new LineBreakEnumerator(textMemory);
+
+            Assert.True(enumerator.MoveNext());
+
+            Assert.Equal(expectedLength, enumerator.Current.PositionWrap);
+        }
+
+        private class LineBreakTestDataGenerator : TestDataGenerator
+        {
+            public LineBreakTestDataGenerator()
+                : base("auxiliary/LineBreakTest.txt")
+            {
+            }
         }
     }
 }
